# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank Cards API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/cards/{id}/request-block": {
      "post": {
        "tags": [
          "Cards"
        ],
        "summary": "Запрос на блокировку карты",
        "description": "Пользователь может запросить блокировку своей карты",
        "operationId": "requestBlock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Запрос обработан / карта заблокирована",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "blocked"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Попытка блокировки чужой карты",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/cards/transfer": {
      "post": {
        "tags": [
          "Cards"
        ],
        "summary": "Перевод между своими картами",
        "description": "Переводит указанную сумму с одной своей карты на другую. Проверяет, что обе карты принадлежат пользователю, активны и достаточно средств.",
        "operationId": "transfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Перевод выполнен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "ok"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неправильный запрос (например, перевод на ту же карту, недостаточно средств)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён (чужая карта)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cards": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Список своих карт",
        "description": "Возвращает страницу собственных карт с фильтрацией по статусу и поиском по последним 4 цифрам или имени владельца",
        "operationId": "listOwn",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы (0-based)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Размер страницы",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Статус карты",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "ACTIVE"
          },
          {
            "name": "search",
            "in": "query",
            "description": "Поисковый термин: последние 4 цифры или имя владельца",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Страница карт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "TransferRequest": {
        "type": "object",
        "description": "Запрос на перевод средств между картами",
        "properties": {
          "fromCardId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID карты-отправителя",
            "example": "e7b9f27a-4c76-4bd1-9f12-123456789abc"
          },
          "toCardId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID карты-получателя",
            "example": "a1b2c3d4-5678-90ab-cdef-1234567890ab"
          },
          "amount": {
            "type": "number",
            "description": "Сумма перевода (положительное число)",
            "example": 1500,
            "minimum": 0.01
          }
        },
        "required": [
          "amount",
          "fromCardId",
          "toCardId"
        ]
      },
...
      "Page": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {

            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}